[feature/MOB-245] - 1- Move the Worker initialisation from TransactionsActivity to App; 2- Remove "onNotificationDismissed" method in SupportNotificationBroadcastReceiver; 3- Refactor constant names in SupportNotificationWorker and change CHANNEL_ID and CHANNEL_NAME to private
