+++ app/src/main/java/com/asfoundation/wallet/support/SupportNotificationWorker.kt
@@ -10,6 +10,9 @@ import androidx.work.RxWorker
 import androidx.work.WorkerParameters
 import com.asf.wallet.R
 import com.asfoundation.wallet.di.WorkerKey
+import com.asfoundation.wallet.support.SupportNotificationBroadcastReceiver.Companion.ACTION_CHECK_MESSAGES
+import com.asfoundation.wallet.support.SupportNotificationBroadcastReceiver.Companion.ACTION_DISMISS
+import com.asfoundation.wallet.support.SupportNotificationBroadcastReceiver.Companion.ACTION_KEY
 import dagger.Binds
 import dagger.Module
 import dagger.multibindings.IntoMap
@@ -31,9 +34,9 @@ class SupportNotificationWorker @Inject constructor(private val context: Context
     const val NOTIFICATION_PERIOD: Long = 15
     const val WORKER_TAG = "SupportNotificationWorkerTag"
     const val UNIQUE_WORKER_NAME = "SupportNotificationWorker"
-    const val channelId = "support_notification_channel_id"
-    const val channelName = "Support notification channel"
     const val UNREAD_CONVERSATIONS = "UNREAD_CONVERSATIONS"
+    private const val CHANNEL_ID = "support_notification_channel_id"
+    private const val CHANNEL_NAME = "Support notification channel"
   }
 
   @Module
@@ -48,8 +51,7 @@ class SupportNotificationWorker @Inject constructor(private val context: Context
     return if (supportInteractor.shouldShowNotification()) {
       supportInteractor.updateUnreadConversations()
       Completable.fromAction {
-            notificationManager.notify(NOTIFICATION_SERVICE_ID,
-                createNotification().build())
+            notificationManager.notify(NOTIFICATION_SERVICE_ID, createNotification().build())
           }
           .andThen(Single.just(Result.success()))
     } else {
@@ -63,12 +65,11 @@ class SupportNotificationWorker @Inject constructor(private val context: Context
     val builder: NotificationCompat.Builder
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
       val importance = NotificationManager.IMPORTANCE_HIGH
-      val notificationChannel =
-          NotificationChannel(channelId, channelName, importance)
-      builder = NotificationCompat.Builder(context, channelId)
+      val notificationChannel = NotificationChannel(CHANNEL_ID, CHANNEL_NAME, importance)
+      builder = NotificationCompat.Builder(context, CHANNEL_ID)
       notificationManager.createNotificationChannel(notificationChannel)
     } else {
-      builder = NotificationCompat.Builder(context, channelId)
+      builder = NotificationCompat.Builder(context, CHANNEL_ID)
     }
     return builder.setContentTitle(context.getString(R.string.support_new_message_title))
         .setAutoCancel(true)
@@ -81,15 +82,13 @@ class SupportNotificationWorker @Inject constructor(private val context: Context
 
   private fun createNotificationClickIntent(): PendingIntent {
     val intent = SupportNotificationBroadcastReceiver.newIntent(context)
-    intent.putExtra(SupportNotificationBroadcastReceiver.ACTION_KEY,
-        SupportNotificationBroadcastReceiver.ACTION_CHECK_MESSAGES)
+    intent.putExtra(ACTION_KEY, ACTION_CHECK_MESSAGES)
     return PendingIntent.getBroadcast(context, 0, intent, 0)
   }
 
   private fun createNotificationDismissIntent(): PendingIntent {
     val intent = SupportNotificationBroadcastReceiver.newIntent(context)
-    intent.putExtra(SupportNotificationBroadcastReceiver.ACTION_KEY,
-        SupportNotificationBroadcastReceiver.ACTION_DISMISS)
+    intent.putExtra(ACTION_KEY, ACTION_DISMISS)
     return PendingIntent.getBroadcast(context, 1, intent, 0)
   }
 }
