+++ app/src/main/java/com/asfoundation/wallet/ui/TransactionsActivity.java
@@ -19,11 +19,6 @@ import androidx.lifecycle.ViewModelProviders;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
-import androidx.work.Constraints;
-import androidx.work.ExistingPeriodicWorkPolicy;
-import androidx.work.NetworkType;
-import androidx.work.PeriodicWorkRequest;
-import androidx.work.WorkManager;
 import com.airbnb.lottie.LottieAnimationView;
 import com.asf.wallet.R;
 import com.asfoundation.wallet.entity.Balance;
@@ -33,7 +28,6 @@ import com.asfoundation.wallet.entity.NetworkInfo;
 import com.asfoundation.wallet.entity.Wallet;
 import com.asfoundation.wallet.referrals.CardNotification;
 import com.asfoundation.wallet.repository.PreferencesRepositoryType;
-import com.asfoundation.wallet.support.SupportNotificationWorker;
 import com.asfoundation.wallet.transactions.Transaction;
 import com.asfoundation.wallet.ui.appcoins.applications.AppcoinsApplication;
 import com.asfoundation.wallet.ui.toolbar.ToolbarArcBackground;
@@ -55,14 +49,10 @@ import dagger.android.AndroidInjection;
 import io.reactivex.Observable;
 import io.reactivex.disposables.CompositeDisposable;
 import io.reactivex.subjects.PublishSubject;
-import java.util.concurrent.TimeUnit;
 import javax.inject.Inject;
 
 import static com.asfoundation.wallet.C.ErrorCode.EMPTY_COLLECTION;
 import static com.asfoundation.wallet.support.SupportNotificationBroadcastReceiver.SUPPORT_NOTIFICATION_CLICK;
-import static com.asfoundation.wallet.support.SupportNotificationWorker.NOTIFICATION_PERIOD;
-import static com.asfoundation.wallet.support.SupportNotificationWorker.UNIQUE_WORKER_NAME;
-import static com.asfoundation.wallet.support.SupportNotificationWorker.WORKER_TAG;
 
 public class TransactionsActivity extends BaseNavigationActivity implements View.OnClickListener {
 
@@ -102,8 +92,6 @@ public class TransactionsActivity extends BaseNavigationActivity implements View
 
     toolbar();
     enableDisplayHomeAsUp();
-
-    setupSupportNotificationWorker(this);
 
     disposables = new CompositeDisposable();
 
@@ -483,19 +471,5 @@ public class TransactionsActivity extends BaseNavigationActivity implements View
     if (showScroll) {
       viewModel.fetchTransactions(false);
     }
-  }
-
-  private void setupSupportNotificationWorker(Context context) {
-    Constraints workerConstraints =
-        new Constraints.Builder().setRequiredNetworkType(NetworkType.CONNECTED)
-            .build();
-    PeriodicWorkRequest notificationWorkRequest =
-        new PeriodicWorkRequest.Builder(SupportNotificationWorker.class, NOTIFICATION_PERIOD,
-            TimeUnit.MINUTES).addTag(WORKER_TAG)
-            .setConstraints(workerConstraints)
-            .build();
-    WorkManager.getInstance(context)
-        .enqueueUniquePeriodicWork(UNIQUE_WORKER_NAME, ExistingPeriodicWorkPolicy.KEEP,
-            notificationWorkRequest);
   }
 }
